API/requirements.txt
fastapi
uvicorn 
motor 
pydantic[dotenv]
numpy
transformers
sentence_transformers
pandas
requests
bs4
python-dotenv
sqlalchemy 
alembic 
openai>=1.0.0
langchain
langchain-openai
asyncio
python-telegram-bot
elasticsearch